Course2/Part2/Python [Практикум]/file manager/requirements.txt
multipledispatch==0.6.0
python-dotenv==0.16.0
six==1.15.0
termcolor==1.1.0
